pyproject.toml
[tool.poetry]
name = "intern-hunter-ai"
version = "0.1.0"
description = "The most advanced, local-first CLI tool for automating high-signal remote AI/ML internships."
authors = ["Kamyavardhan Dave <kamyavardhan@example.com>"]
readme = "README.md"
packages = [{include = "intern_hunter", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
playwright = "^1.41.0"
google-generativeai = "^0.4.0"
groq = "^0.4.2"
weasyprint = "^61.0"
Jinja2 = "^3.1.3"
sendgrid = "^6.11.0"
python-telegram-bot = "^20.7"
PyYAML = "^6.0.1"
PyMuPDF = "^1.23.21"
apscheduler = "^3.10.4"
rich = "^13.7.0"
bs4 = "^0.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
intern-hunter = "intern_hunter.cli:app"
